Extract shared siteadmin AdminAPIService
