+++ pkg/siteadmin/service/app.go
@@ -4,22 +4,16 @@ import (
 	"context"
 	"database/sql"
 	"errors"
-	"fmt"
-	"net/http"
 	"strings"
 
 	sq "github.com/Masterminds/squirrel"
 
 	"github.com/authgear/authgear-server/pkg/api/siteadmin"
-	relay "github.com/authgear/authgear-server/pkg/graphqlgo/relay"
 	"github.com/authgear/authgear-server/pkg/lib/analytic"
 	"github.com/authgear/authgear-server/pkg/lib/config/configsource"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db/auditdb"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db/globaldb"
-	portalservice "github.com/authgear/authgear-server/pkg/portal/service"
-	"github.com/authgear/authgear-server/pkg/portal/session"
 	"github.com/authgear/authgear-server/pkg/util/clock"
-	"github.com/authgear/authgear-server/pkg/util/graphqlutil"
 	"github.com/authgear/authgear-server/pkg/util/timeutil"
 )
 
@@ -43,14 +37,6 @@ type AppServiceOwnerStore interface {
 	GetOwnersByAppIDs(ctx context.Context, appIDs []string) (map[string]string, error)
 	CountAppsByOwnerUserID(ctx context.Context, userID string) (int, error)
 	ListAppIDsByOwnerUserIDPaged(ctx context.Context, userID string, limit uint64, offset uint64) ([]string, error)
-}
-
-type AppServiceAdminAPI interface {
-	SelfDirector(ctx context.Context, actorUserID string, usage portalservice.Usage) (func(*http.Request), error)
-}
-
-type AppServiceHTTPClient struct {
-	*http.Client
 }
 
 // ---- AppOwnerStore -----------------------------------------------------------
@@ -175,10 +161,9 @@ type AppService struct {
 	GlobalDatabase    AppServiceDatabase
 	ConfigSourceStore AppServiceConfigSourceStore
 	OwnerStore        AppServiceOwnerStore
-	AdminAPI          AppServiceAdminAPI
+	AdminAPI          *AdminAPIService
 	AuditDatabase     *auditdb.ReadHandle
 	AuditStore        *analytic.AuditDBReadStore
-	HTTPClient        AppServiceHTTPClient
 	Clock             clock.Clock
 }
 
@@ -429,130 +414,12 @@ func (s *AppService) fetchTotalUserCount(ctx context.Context, appID string) (int
 // findUserIDsByEmail calls Admin API getUsersByStandardAttribute to find users
 // matching the given email. Returns their raw (non-global) user IDs.
 func (s *AppService) findUserIDsByEmail(ctx context.Context, email string) ([]string, error) {
-	params := graphqlutil.DoParams{
-		OperationName: "getUsersByStandardAttribute",
-		Query: `
-		query getUsersByStandardAttribute($name: String!, $value: String!) {
-			users: getUsersByStandardAttribute(attributeName: $name, attributeValue: $value) {
-				id
-			}
-		}
-		`,
-		Variables: map[string]interface{}{
-			"name":  "email",
-			"value": email,
-		},
-	}
-
-	r, err := http.NewRequestWithContext(ctx, "POST", "/graphql", nil)
-	if err != nil {
-		return nil, err
-	}
-
-	actorUserID := session.GetValidSessionInfo(ctx).UserID
-	director, err := s.AdminAPI.SelfDirector(ctx, actorUserID, portalservice.UsageInternal)
-	if err != nil {
-		return nil, err
-	}
-	director(r)
-
-	result, err := graphqlutil.HTTPDo(s.HTTPClient.Client, r, params)
-	if err != nil {
-		return nil, err
-	}
-	if result.HasErrors() {
-		return nil, fmt.Errorf("failed to search users by email: %v", result.Errors)
-	}
-
-	data := result.Data.(map[string]interface{})
-	users := data["users"].([]interface{})
-
-	ids := make([]string, 0, len(users))
-	for _, u := range users {
-		userNode, ok := u.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		globalID, _ := userNode["id"].(string)
-		resolved := relay.FromGlobalID(globalID)
-		if resolved == nil || resolved.ID == "" {
-			// relay.FromGlobalID failed to parse — skip this entry.
-			continue
-		}
-		ids = append(ids, resolved.ID)
-	}
-	return ids, nil
+	return s.AdminAPI.FindUserIDsByEmail(ctx, email)
 }
 
 // resolveUserEmails batch-fetches emails for the given user IDs via Admin API.
 func (s *AppService) resolveUserEmails(ctx context.Context, userIDs []string) (map[string]string, error) {
-	if len(userIDs) == 0 {
-		return map[string]string{}, nil
-	}
-
-	globalIDs := make([]string, len(userIDs))
-	for i, id := range userIDs {
-		globalIDs[i] = relay.ToGlobalID("User", id)
-	}
-
-	params := graphqlutil.DoParams{
-		OperationName: "getUserNodes",
-		Query: `
-		query getUserNodes($ids: [ID!]!) {
-			nodes(ids: $ids) {
-				... on User {
-					id
-					standardAttributes
-				}
-			}
-		}
-		`,
-		Variables: map[string]interface{}{
-			"ids": globalIDs,
-		},
-	}
-
-	r, err := http.NewRequestWithContext(ctx, "POST", "/graphql", nil)
-	if err != nil {
-		return nil, err
-	}
-
-	actorUserID := session.GetValidSessionInfo(ctx).UserID
-	director, err := s.AdminAPI.SelfDirector(ctx, actorUserID, portalservice.UsageInternal)
-	if err != nil {
-		return nil, err
-	}
-	director(r)
-
-	result, err := graphqlutil.HTTPDo(s.HTTPClient.Client, r, params)
-	if err != nil {
-		return nil, err
-	}
-	if result.HasErrors() {
-		return nil, fmt.Errorf("failed to resolve user emails: %v", result.Errors)
-	}
-
-	emailMap := make(map[string]string, len(userIDs))
-	data := result.Data.(map[string]interface{})
-	nodes := data["nodes"].([]interface{})
-	for _, node := range nodes {
-		userNode, ok := node.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		globalID, _ := userNode["id"].(string)
-		resolvedID := relay.FromGlobalID(globalID)
-		if resolvedID == nil || resolvedID.ID == "" {
-			continue
-		}
-		attrs, ok := userNode["standardAttributes"].(map[string]interface{})
-		if !ok {
-			continue
-		}
-		email, _ := attrs["email"].(string)
-		emailMap[resolvedID.ID] = email
-	}
-	return emailMap, nil
+	return s.AdminAPI.ResolveUserEmails(ctx, userIDs)
 }
 
 func uniqueValues(m map[string]string) []string {
