+++ pkg/siteadmin/service/deps.go
@@ -8,13 +8,14 @@ import (
 	"github.com/authgear/authgear-server/pkg/util/httputil"
 )
 
-func NewHTTPClient() AppServiceHTTPClient {
-	return AppServiceHTTPClient{Client: httputil.NewExternalClient(5 * time.Second)}
+func NewHTTPClient() SiteAdminHTTPClient {
+	return SiteAdminHTTPClient{Client: httputil.NewExternalClient(5 * time.Second)}
 }
 
 var DependencySet = wire.NewSet(
 	wire.Struct(new(AppOwnerStore), "*"),
 	wire.Bind(new(AppServiceOwnerStore), new(*AppOwnerStore)),
+	wire.Struct(new(AdminAPIService), "*"),
 	wire.Struct(new(AppService), "*"),
 	NewHTTPClient,
 )
